chore: move chat component
