+++ src/components/chat/index.tsx
@@ -1,8 +1,11 @@
+import { useEffect } from 'react';
+
 import useSendbirdStateContext from '@uikit/hooks/useSendbirdStateContext';
 
 import { ChatContainer } from './context/ChatProvider';
 import { ChatUI } from './ui';
-import { useWidgetSetting } from '../../context/WidgetSettingContext';
+import { useWidgetSession, useWidgetSetting } from '../../context/WidgetSettingContext';
+import useWidgetButtonActivityTimeout from '../../hooks/useWidgetButtonActivityTimeout';
 
 export const WidgetChatting = () => {
   const { stores } = useSendbirdStateContext();
@@ -22,3 +25,29 @@ export const WidgetChatting = () => {
     </ChatContainer>
   );
 };
+
+const Chat = ({ fullscreen = false }: { fullscreen?: boolean }) => {
+  useWidgetButtonActivityTimeout(fullscreen);
+  const { stores } = useSendbirdStateContext();
+  const widgetSetting = useWidgetSetting();
+  const widgetSession = useWidgetSession();
+
+  // Initialize the manual session if channelUrl is not set.
+  useEffect(() => {
+    if (widgetSetting.initialized && stores.sdkStore.initialized) {
+      if (widgetSession.strategy === 'manual' && !widgetSession.channelUrl) {
+        widgetSetting.initManualSession(stores.sdkStore.sdk);
+      }
+    }
+  }, [
+    widgetSetting.initialized,
+    widgetSession.strategy,
+    widgetSession.channelUrl,
+    stores.sdkStore.sdk,
+    stores.sdkStore.initialized,
+  ]);
+
+  return <WidgetChatting />;
+};
+
+export default Chat;
