+++ src/components/widget/ChatAiWidget.tsx
@@ -2,7 +2,6 @@ import styled from 'styled-components';
 
 import { StringSet } from '@uikit/ui/Label/stringSet';
 
-import Chat from './Chat';
 import ProviderContainer from './ProviderContainer';
 import WidgetToggleButton from './WidgetToggleButton';
 import WidgetWindow from './WidgetWindow';
@@ -11,6 +10,7 @@ import { useWidgetState } from '../../context/WidgetStateContext';
 import useMobileView from '../../hooks/useMobileView';
 import { useWidgetAutoOpen } from '../../hooks/useWidgetAutoOpen';
 import { isMobile } from '../../utils';
+import Chat from '../chat';
 
 const MobileContainer = styled.div<{ width: number }>`
   position: fixed;
